+++ src/MapLoader/MapObject.java
@@ -3,13 +3,13 @@ package MapLoader;
 import java.util.Map;
 
 public class MapObject {
-	public int x;
-	public int y;
-	public int height;
-	public int width;
+	public double x;
+	public double y;
+	public double height;
+	public double width;
 	public Map<String, String> properties = null;
 
-	public MapObject(int x, int y, int height, int width, Map<String, String> properties) {
+	public MapObject(double x, double y, double height, double width, Map<String, String> properties) {
 		this.x=x;
 		this.y=y;
 		this.height=height;
